+++ docs/changes.rst
@@ -6,6 +6,11 @@ Version 0.2.0
 
 To be released.
 
+- Added :meth:`SubscriptionSet.contains()
+  <libearth.subscribe.SubscriptionSet.contains>` method which provides
+  ``recursively=True`` option.  It's useful for determining that
+  a subcategory or subscription is in the whole tree.
+
 
 Version 0.1.0
 -------------